Initial Kibana service implementation
Kibana is a Node.js application that uses Elasticsearch as its upstream.
We configure it with Containerbuddy to have an `onStart` and `onChange`
handler that rewrites its config file to find the ES master-only nodes.

The ES compose files are pulled from the `triton-elasticsearch` blueprint
and self-configure in the same way that stack does.